requirements.txt
opencv-python
numpy
pyyaml
redis[hiredis]
jinja-compose-wrapper